+++ apps/router/src/gateway-ui/components/walletModal/send/SendEcash.tsx
@@ -4,12 +4,11 @@ import { useTranslation } from '@fedimint/utils';
 import { Sats } from '@fedimint/types';
 import FederationSelector from '../FederationSelector';
 import { AmountInput, CreateButton, QRCodeTabs } from '..';
-import { useGatewayApi, useGatewayContext } from '../../../../context/hooks';
+import { useGatewayContext } from '../../../../hooks';
 
 const SendEcash: React.FC = () => {
   const { t } = useTranslation();
-  const { state } = useGatewayContext();
-  const api = useGatewayApi();
+  const { state, api } = useGatewayContext();
   const [amount, setAmount] = useState<Sats>(0 as Sats);
   const [ecash, setEcash] = useState<string>('');
   const [showEcash, setShowEcash] = useState<boolean>(false);
